Private headers for connector
