+++ src/connection_manager.c
@@ -24,6 +24,7 @@
 #include <qpid/dispatch/failoverlist.h>
 #include <proton/listener.h>
 #include "connection_private.h"
+#include "connector_private.h"
 #include "dispatch_private.h"
 #include "connection_manager_private.h"
 #include "server_private.h"
