+++ src/policy.c
@@ -24,6 +24,7 @@
 #include <stdio.h>
 #include <string.h>
 #include "connection_private.h"
+#include "connector_private.h"
 #include "dispatch_private.h"
 #include "qpid/dispatch/container.h"
 #include "qpid/dispatch/server.h"
