+++ src/server_private.h
@@ -44,9 +44,6 @@ qd_connection_t *qd_server_connection(qd_server_t *server, qd_server_config_t* c
 qd_listener_t *qd_server_listener(qd_server_t *server);
 qd_connector_t *qd_server_connector(qd_server_t *server);
 
-const qd_server_config_t *qd_connector_config(const qd_connector_t *c);
-
-bool qd_connector_decref(qd_connector_t* ct);
 void qd_listener_decref(qd_listener_t* ct);
 void qd_server_config_free(qd_server_config_t *cf);
 
@@ -97,41 +94,8 @@ struct qd_listener_t {
 
 DEQ_DECLARE(qd_listener_t, qd_listener_list_t);
 
-/**
- * Connector objects represent the desire to create and maintain an outgoing transport connection.
- */
-struct qd_connector_t {
-    /* May be referenced by connection_manager, timer and pn_connection_t */
-    sys_atomic_t              ref_count;
-    qd_server_t              *server;
-    qd_server_config_t        config;
-    qd_timer_t               *timer;
-    long                      delay;
-
-    /* Connector state and ctx can be modified in proactor or management threads. */
-    sys_mutex_t              *lock;
-    cxtr_state_t              state;
-    char                     *conn_msg;
-    qd_connection_t          *ctx;
-
-    /* This conn_list contains all the connection information needed to make a connection. It also includes failover connection information */
-    qd_failover_item_list_t   conn_info_list;
-    int                       conn_index; // Which connection in the connection list to connect to next.
-
-    /* Optional policy vhost name */
-    char                     *policy_vhost;
-
-    DEQ_LINKS(qd_connector_t);
-};
-
-DEQ_DECLARE(qd_connector_t, qd_connector_list_t);
-
-const char *qd_connector_policy_vhost(qd_connector_t* ct);
-
-
 ALLOC_DECLARE(qd_listener_t);
 ALLOC_DECLARE(qd_deferred_call_t);
-ALLOC_DECLARE(qd_connector_t);
 ALLOC_DECLARE(qd_pn_free_link_session_t);
 
 DEQ_DECLARE(qd_connection_t, qd_connection_list_t); // XXX JRR
@@ -164,8 +128,6 @@ struct qd_server_t {
  * For every connection on the server's connection list, call pn_transport_set_tracer and enable trace logging
  */
 void qd_server_trace_all_connections();
-
-qd_failover_item_t *qd_connector_get_conn_info(qd_connector_t *connector);
 
 /**
  * This function is set as the pn_transport->tracer and is invoked when proton tries to write the log message to pn_transport->tracer
